dirmodel: handle symlinks
In case of symlinks, stat the pointed to file but show, that it was a
symlink. This also consolidates the reading of the filedata for the
initial population of the dirmodel and the files added through inotify.
As the inotify handler has already resolved symlinks accidentally this
inconsistency is now resolved.