listview: handle file addition to an empty dir in change_cb
If a directory was empty and a file was added, the index got a value of
1, thereby being out of bounds. The only case, where it was not like
that, was, when the directory was entered before and than emptied.
But this was only a side effect of a integer underflow during the
removal of files.

To fix the issue, just set the index to 0 if the directory contains just
one file.