dirmodel: make it work for paths other than cwd
Although it was possible to specify other directories during init or
change_directory, it only really worked for the current working directory, as
stat calls where all relative to cwd and the inotify watcher was registered for
".".

Now the DIR handle from init is kept open and the corresponding fd is used in
fstatat calls to make them independent of the cwd and the inotify watcher is
registered for the provided path.