dirmodel: respect wcwidth of each char during render
As several unicode characters only modify the previous character,
thereby using no screen space, or are displayed in two screen columns,
one cannot just count the characters to determine the needed columns.

The dirmodel_render function and its helper functions now use
wc(s)width, to determine the real needed space. In case the supplied
render buffer is to small, it stops rendering but returns the needed
buffer size.

A small modification was needed to listmodel, to reflect the render
function signature change and listview now respects the return value of
listview_render to adjust the buffer size and rerender the buffer.