main: allow the file open handler to handle marked files
Instead giving the selected file as a parameter to the file open
handler, a temporary directory is now created and the selected file is
stored in the file "selected". Additionally, marked files are stored in
the file "marked" separated by a null character. The temporary directory
is then given to the open handler as the first parameter.

An example open handler script is also provided. It opens the selected
file, if it is not one of the marked files. Otherwise it opens all
marked files.