test: add testing for OOM situations
A wrapper for malloc, realloc and strdup are added which can be
programmed to fail after a predefined number of invocations. Each test
is now run multiple time, starting with an instant fail of these
functions and increasing the not failing invocations on every pass. This
is to ensure, that all code paths using allocation functions handle OOM
correctly.

Tests have to use the new assert_oom macro on conditions, that fail
during OOM. When not testing OOM it expands to ck_assert. When testing
OOM it negates the condition and in case it is true, it signals the main
process about it via a pipe and returns from the test. This is used by
the main process to decide, when testing is finished (no test hit an OOM
condition). The macro can only be used in the test body, not in a setup
function as it cannot abort a test in setup without making the test
fail.

The END_TEST macro of check is also overriden to make the behaviour of
the allocation functions normal again. This is especially needed,
because the coverage functions also need a working malloc, which would
crash after the test. The assert_oom macro also reset the behaviour in
case of OOM, as it is another exit point of a test.

Currently all tests fail, as no OOM handling is implemented in neither
the tested modules nor the tests.