dict, tests: fix OOM handling
The dict tests still leaked memory and had illegal frees. The test data
is now duplicated with the real strdup function instead of the failing
on, to ease testing. There is now also a new assert_oom_cleanup macro,
with executes all expressions following the condition in case the
condition is not true during OOM testing.

Addtionally, the dict_set function now only allocates memory, if it
really inserts something into the dictionary.