dirmodel: return error codes in dirmodel_getmarkedfiles
The error codes designate if the list could not be created because of an
out of memory situation or because no file was marked.

The file list is now an out parameter. It is made const, because all
previous usages also use a const list_t pointer. If it is later needed
non-const, the const modifier could be removed.