application: fix two issues in invoke_handler
The first issues was that for processes spawned in the background, the
return value check was inverted, thereby trying to remove the temp
directory. It just did not happen because of the secound issue: the code
tried to remove the directory with the remove function, which only works
for empty directories. This is now done with remove_directory_recursively.