Implement shared clipboard
Multiple instances of dfm can now share the clipboard. The shared
clipboard is a directory specified in the environment variable
DFM_CLIPBOARD_DIR. It must exist, before dfm is started. Otherwise dfm
falls back to the non-shared clipboard.

The implementation is currently not race-free, but as dfm is an
interactive application, one might not be able to trigger any problems.
However, if it turns out, that it is a problem, a race-free variant can
be implemented.