keymap: make use of struct list instead of custom dynamic memory management
Also, setfromstring won't call keymap_destroy anymore, so it is now the
obligation of the callee to destroy the keymap after a failure. This
also prevents a potential double free, as the application object always
called keymap_destroy even when it could already be destroyed because of
a keymap loading error.