listmodel, listview: support moved elements in MODEL_CHANGE
This is needed when changing a list item results in moving it to another
position, e.g. because of sort order.

The listmodel change event had to be extended for that. So all call
sites and tests had to be fixed accordingly.

Additionally, a missing break statement is added to the listview change
callback. Before adding this new support, it resulted only in a
performance impact (printing the list twice when removing an item in
some cases), but now it would change behaviour/break tests.