dirmodel: prepare for supporting different sorting routines
There are now two lists managing the filedata structs. The old one
(list) is used for fast lookup of filenames. The new one (sortedlist) is
used to implement the listmodel interface, thereby visible in the
listview. The dirmodel now also fully supports if a file changes its
position in the sortedlist due to attribute changes.