application, listview: throttle screen updates for inotify events
Actions which would previously redraw the listview now only set a flag,
that the listview needed to be refreshed. For actions triggered by the
user, the application refreshes the listview immediately. For updates
that were triggered by inotify events, the application just starts a
timer of 50ms and refreshes the screen, when the timer is expired. This
reduces the cpu load significantly when a lot of inotify events arrive.
Up to 30% load can be saved with this optimization.