dirmodel: queue add or change events and handle them later
For every file add or change event, the dirmodel stats the file through
the filedata_new_from_file and looks up the file in two lists using
complex comparision functions. This is relatively time consuming.  If a
file changes a lot, this can result in a lot of repeated executions
meaning there is a high cpu load. At the moment, the screen is refreshed
only every 50ms at max, so keeping all internal data structurs up to
date is unecessary.

If the dirmodel now receives an add or change event, it just stores the
file name thereby merging multiple events for a file into a single
event. The dirmodel data structures are then updated when the
inotify_flush function is called.

In a test case this resulted in a CPU load reduction by about 70%.
Strangely, we now spent more then 10 times as much time in system call
context in this test case. That is two thirds of the overall load, where
it previously was negligible. I assume the events our now handled so
fast, that they don't queue up in the file descriptor and they are
handled one by one, meaning more calls to poll and read.