dirmodel: only stop flushing events in OOM situations
When multiple files had add or change events and one of the files where
deleted before flushing, an ENOENT error was returned when trying to
really add the file. As the loop stopped processing the files on every
error, all further files in the queue where not added or updated. The
notify_flush function now only stops on ENOMEM and also only returns
ENOMEM as an error, as ENOENT is expected behaviour.