tests: try to use the current locale instead of forcing en_US.UTF-8
Some tests need an UTF-8 locale to function properly. As it isn't known
if the system uses a UTF-8 locale, en_US.UTF-8 was forced during tests.
This would fail all UTF-8 related tests, if the system didn't support
this locale. As it doesn't matter which locale is used, as long as it is
an UTF-8 locale, just use the currently set locale and test, if it for
the correct encoding. If it isn't an UTF-8 locale, we still fall back to
forcing en_US.UTF-8, and if this also fails, we abort the whole test
run.