examples: reset RUNNING_IN_XTERM in xterminal script
When one launches dfm itself in a terminal previously open by dfm, the
RUNNING_IN_XTERM environment variable is already set to "t". This means,
that all handlers launched in the new dfm instance and are using the
xterminal script, don't open a new terminal. Resetting the variable
at the end of the terminal script (here, we are already in the launched
terminal) fixes this.

Only the shell handler was affected, but by fixing this in the xterminal
script, other handlers which are added later and could be affected are
also fixed.