application: fix selected new file name after rename command
The rename command originally selected the renamed file after the rename
operation. After some performance improvements in the dirmodel
(asynchronous handling of file change events), the new name was not
contained in the dirmodel when trying to select it, thus another file
was selected. After rename, we now flush the file change queue to be
able to select the file under the new name.

To further reduce flicker (for a short time, both file names are
visible), the file is now also deleted from the model in the rename
function.

Also, only refresh the statusbar in NORMAL mode.