commandline: initialize buffer correctly after using the history
When the history function was used, but the command was not edited, the
buffer pointer still pointed to the history entry on the next command
line invocation and not the editable command buffer. This could result
in mutiple memory errors, e.g. overwriting the history entry, buffer
overflow and double free.