Handle files with illegal characters better
When displaying filenames, instead of first completely converting
filenames to wchar_t with mbstowcs and replacing the whole filename with
"???" in case there was a conversion error, convert it manually
character by character with mbtowc and replace unconvertible bytes with
the Unicode replacement character. This makes the filename still
readable in case the encoding of the filename overlaps in parts the with
current locale.